+++ cpp/src/serialize.cpp
@@ -69,7 +69,7 @@ string State::to_json() {
 
         // print coordinates of precinct
         str += TAB(4) + OQ + "coordinates" + CQC + "[";
-        for (vector<int> coordset: district.border) {
+        for (vector<float> coordset : district.border) {
             str += "[" + RED + to_string(coordset[0]) + RESET + ", " 
                 + RED + to_string(coordset[1]) + RESET + "], ";
         }
